+++ src/settingsDialog.hpp
@@ -49,6 +49,7 @@ private:
 	Q_SLOT void on_alwaysShowArrowEndConnectionMatrixCheckBox_toggled(bool checked);
 
 	// Controller
+	Q_SLOT void on_discoveryDelayLineEdit_returnPressed();
 	Q_SLOT void on_enableAEMCacheCheckBox_toggled(bool checked);
 	Q_SLOT void on_fullAEMEnumerationCheckBox_toggled(bool checked);